updated variable names, functions etc.
