+++ src/providers/AlertProvider.jsx
@@ -4,6 +4,11 @@ import AlertContext from '@/contexts/AlertContext';
 import AlertBox from '@/components/AlertBox';
 import { useState } from 'react';
 
+/**
+ * Custom Alert Provider for alert boxes
+ * @param {JSX} children
+ * @returns 
+ */
 const AlertProvider = ({ children }) => {
     const [alertMessage, setAlertMessage] = useState('');
     const [alertType, setAlertType] = useState('');
