+++ src/providers/CalendarDataProvider.jsx
@@ -82,77 +82,3 @@ export const CalendarDataProvider = ({ children }) => {
         </CalendarDataContext.Provider>
     );
 };
-
-    // const { session, userRole } = useAuthSession();
-    // const userEmail = session?.user?.email;
-    // const [allEvents, setAllEvents] = useState([]);
-    // const [availabilities, setAvailabilities] = useState([]);
-    // const [studentRequests, setStudentRequests] = useState([]);
-    // const [subjects, setSubjects] = useState([]);
-    // const [tutors, setTutors] = useState([]);
-    // const [students, setStudents] = useState([]);
-    // const [loading, setLoading] = useState(true);
-
-    // // Fetch all calendar data
-    // useEffect(() => {
-    //     if (!userRole || !userEmail) return;
-
-    //     setLoading(true);
-    //     firestoreFetchShifts(
-    //         userRole,
-    //         userEmail,
-    //         setAllEvents,
-    //         setAllEvents,
-    //         setStudents,
-    //         setLoading,
-    //     );
-    //     firestoreFetchAvailabilities(setAvailabilities);
-    //     firestoreFetchStudentRequests(setStudentRequests);
-    //     firestoreFetchSubjectsWithTutors(setSubjects);
-    //     firestoreFetchTutors(setTutors);
-    // }, [userRole, userEmail]);
-
-    // // Split availabilities based on events (memoized)
-    // const splitAvailabilitiesData = useMemo(
-    //     () => calendarAvailabilitySplit(availabilities, allEvents),
-    //     [availabilities, allEvents],
-    // );
-
-    // // Email queue monitoring (side effect for teachers)
-    // useEmailQueueMonitor(userRole);
-
-    // const value = useMemo(
-    //     () => ({
-    //         allEvents,
-    //         setAllEvents,
-    //         availabilities,
-    //         setAvailabilities,
-    //         studentRequests,
-    //         setStudentRequests,
-    //         splitAvailabilitiesData,
-    //         subjects,
-    //         tutors,
-    //         students,
-    //         loading,
-    //         userRole,
-    //         userEmail,
-    //     }),
-    //     [
-    //         allEvents,
-    //         availabilities,
-    //         studentRequests,
-    //         splitAvailabilitiesData,
-    //         subjects,
-    //         tutors,
-    //         students,
-    //         loading,
-    //         userRole,
-    //         userEmail,
-    //     ],
-    // );
-
-    // return (
-	// 	<CalendarDataContext.Provider value={value}>
-	// 		{children}
-	// 	</CalendarDataContext.Provider>
-	// );
