+++ src/utils/msTeams.js
@@ -180,7 +180,8 @@ export const updateTeamsMeeting = async (eventId, subject, description, startTim
             const startDate = new Date(startTime);
             const endDate = recurrenceOptions.until || new Date(startDate);
             if (!recurrenceOptions.until) {
-                endDate.setMonth(endDate.getMonth() + 3); // Default 3 months
+                endDate.setMonth(endDate.getMonth() + 3); 
+                // 3 months default
             }
 
             eventBody.recurrence = {
